@@ -1,7 +1,7 @@
 [package]
 name = "siquery_cli"
 version = "1.2.5"
-edition = "2018"
+edition = "2021"
 readme = "../README.md"
 license = "MIT/Apache-2.0"
 homepage = "https://github.com/devolutions/siquery-rs"
