@@ -42,7 +42,7 @@ impl SystemInfoData {
         }
     }
 
-    pub(crate) fn get_specific_ex(reader: &SystemInfoDataIface) -> Vec<SystemInfoData> {
+    pub(crate) fn get_specific_ex(reader: &dyn SystemInfoDataIface) -> Vec<SystemInfoData> {
         let mut output : Vec<SystemInfoData> = Vec::new();
         let mut system_info = SystemInfoData::new();
         system_info.computer_name = reader.hostname().unwrap_or(String::from(""));
@@ -51,7 +51,7 @@ impl SystemInfoData {
     }
 
     pub(crate) fn get_specific() -> Vec<SystemInfoData> {
-        let reader: Box<SystemInfoDataIface> = Box::new(Reader{});
+        let reader: Box<dyn SystemInfoDataIface> = Box::new(Reader{});
         let out = SystemInfoData::get_specific_ex(reader.borrow());
         out
     }
@@ -83,7 +83,7 @@ mod tests {
     }
     #[test]
     fn test_system_info () {
-        let system_reader: Box<SystemInfoDataIface> = Box::new(Test{});
+        let system_reader: Box<dyn SystemInfoDataIface> = Box::new(Test{});
         let system_info = &SystemInfoData::get_specific_ex(system_reader.borrow())[0];
         assert_eq!(system_info.computer_name, "galaxy500");
         assert_eq!(system_info.cpu_logical_cores, 0);
